Add server-side apply options
This adds patch options related to server-side apply, and moves options
to their own file.
